size_t: update bitmap lib
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@52042 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/bitmap/bitmap.c
@@ -242,12 +242,12 @@ int BM_get(struct BM *map, int x, int y)
  *  \return int
  */
 
-int BM_get_map_size(struct BM *map)
+size_t BM_get_map_size(struct BM *map)
 {
     if (map->sparse)
 	return BM_get_map_size_sparse(map);
 
-    return map->bytes * map->rows;
+    return (size_t) map->bytes * map->rows;
 }
